Use a custom timing to adjust the planetarium camera.
