+++ ksp_plugin_adapter/ksp_plugin_adapter.csproj
@@ -125,6 +125,7 @@
     <Compile Include="main_window.cs" />
     <Compile Include="map_node_pool.cs" />
     <Compile Include="orbit_analyser.cs" />
+    <Compile Include="planetarium_camera_adjuster.cs" />
     <Compile Include="rendering_actions.cs" />
     <Compile Include="null_extensions.cs" />
     <Compile Include="optional_marshaler.cs" />
